refactor: folders structure
