+++ lib/app/features/wallet/views/pages/coins_flow/network_list/network_item.dart
